Modifying an item now takes time and stamina
Modifying includes, so far, equipping, unequipping, and dropping the item. If the actor doesn't have enough stamina, the action cannot be completed. During the implementation of this update, a bug was fixed that was related to setting a mutable object as a parameter to a keyword argument (in mobs.py, buffs=[] and multipliers=[] created the bug.).